Face height

+++ main.py
@@ -68,7 +68,7 @@ for face_landmarks in face_landmarks_list:
     "## Head position"
     chin_template = Image.open("Kinnschablone.png").convert("RGBA")
     height = lower - upper
-    top_to_chin = h * 1.2 + h
+    top_to_chin = int(h * 1.2 + h)
     offset = chin_template.height - top_to_chin
     bordered_image = Image.new('RGBA', (img.width, img.height + 200), (255, 0, 0, 0))
     bordered_image.paste(equalized, (0, 200))
